feat: move regen into BeaconEngine
