+++ packages/beacon-node/src/chain/beaconEngine/beaconEngine.ts
@@ -1,6 +1,11 @@
 import {BeaconConfig} from "@lodestar/config";
-import {IBeaconStateView} from "@lodestar/state-transition";
+import {CheckpointWithHex, ForkChoiceStateGetter, IForkChoice} from "@lodestar/fork-choice";
+import {EffectiveBalanceIncrements, IBeaconStateView, computeEpochAtSlot} from "@lodestar/state-transition";
+import {Logger, toRootHex} from "@lodestar/utils";
+import {Metrics} from "../../metrics/index.js";
+import {CheckpointBalancesCache} from "../balancesCache.js";
 import {BlsMultiThreadWorkerPool, BlsSingleThreadVerifier, IBlsVerifier} from "../bls/index.js";
+import {initializeForkChoice} from "../forkChoice/index.js";
 import {
   AggregatedAttestationPool,
   AttestationPool,
@@ -9,6 +14,7 @@ import {
   SyncCommitteeMessagePool,
   SyncContributionAndProofPool,
 } from "../opPools/index.js";
+import {QueuedStateRegenerator} from "../regen/index.js";
 import {
   SeenAggregators,
   SeenAttesters,
@@ -18,6 +24,9 @@ import {
 } from "../seenCache/index.js";
 import {SeenAttestationDatas} from "../seenCache/seenAttestationData.js";
 import {ShufflingCache} from "../shufflingCache.js";
+import {FIFOBlockStateCache} from "../stateCache/fifoBlockStateCache.js";
+import {PersistentCheckpointStateCache} from "../stateCache/persistentCheckpointsCache.js";
+import {BlockStateCache, CheckpointStateCache} from "../stateCache/types.js";
 import {BeaconEngineModules, IBeaconEngine} from "./interface.js";
 
 /**
@@ -28,8 +37,16 @@ import {BeaconEngineModules, IBeaconEngine} from "./interface.js";
  */
 export class BeaconEngine implements IBeaconEngine {
   readonly config: BeaconConfig;
+  private readonly logger: Logger;
+  private readonly metrics: Metrics | null;
   readonly bls: IBlsVerifier;
   readonly shufflingCache: ShufflingCache;
+
+  readonly blockStateCache: BlockStateCache;
+  readonly checkpointStateCache: CheckpointStateCache;
+  readonly checkpointBalancesCache: CheckpointBalancesCache;
+  readonly forkChoice: IForkChoice;
+  readonly regen: QueuedStateRegenerator;
 
   // Op pools
   readonly attestationPool: AttestationPool;
@@ -48,8 +65,25 @@ export class BeaconEngine implements IBeaconEngine {
   readonly seenAttestationDatas: SeenAttestationDatas;
 
   constructor(modules: BeaconEngineModules, anchorState: IBeaconStateView) {
-    const {opts, config, logger, metrics, clock, pubkeyCache} = modules;
+    const {
+      opts,
+      config,
+      logger,
+      metrics,
+      clock,
+      pubkeyCache,
+      bufferPool,
+      cpStateDatastore,
+      emitter,
+      signal,
+      db,
+      validatorMonitor,
+      seenBlockInputCache,
+      isAnchorStateFinalized,
+    } = modules;
     this.config = config;
+    this.logger = logger;
+    this.metrics = metrics;
 
     // by default, verify signatures on both main threads and worker threads
     this.bls = opts.blsVerifyAllMainThread
@@ -80,5 +114,142 @@ export class BeaconEngine implements IBeaconEngine {
 
     this.seenContributionAndProof = new SeenContributionAndProof(metrics);
     this.seenAttestationDatas = new SeenAttestationDatas(metrics, opts?.attDataCacheSlotDistance);
+
+    this.blockStateCache = new FIFOBlockStateCache(opts, {metrics});
+    this.checkpointStateCache = new PersistentCheckpointStateCache(
+      {config, metrics, logger, clock, blockStateCache: this.blockStateCache, bufferPool, datastore: cpStateDatastore},
+      opts
+    );
+
+    const {checkpoint} = anchorState.computeAnchorCheckpoint();
+    this.blockStateCache.add(anchorState);
+    this.blockStateCache.setHeadState(anchorState);
+    this.checkpointStateCache.add(checkpoint, anchorState);
+
+    this.checkpointBalancesCache = new CheckpointBalancesCache();
+
+    const forkChoiceStateGetter: ForkChoiceStateGetter = ({stateRoot, checkpoint}) => {
+      if (stateRoot) return this.blockStateCache.get(stateRoot);
+
+      if (checkpoint) return this.checkpointStateCache.get({epoch: checkpoint.epoch, rootHex: checkpoint.rootHex});
+
+      return null;
+    };
+
+    this.forkChoice = initializeForkChoice(
+      config,
+      emitter,
+      clock.currentSlot,
+      anchorState,
+      isAnchorStateFinalized,
+      opts,
+      this.justifiedBalancesGetter.bind(this),
+      forkChoiceStateGetter,
+      metrics,
+      logger
+    );
+
+    this.regen = new QueuedStateRegenerator({
+      config,
+      forkChoice: this.forkChoice,
+      blockStateCache: this.blockStateCache,
+      checkpointStateCache: this.checkpointStateCache,
+      seenBlockInputCache,
+      db,
+      metrics,
+      validatorMonitor,
+      logger,
+      emitter,
+      signal,
+    });
+  }
+
+  /**
+   * `ForkChoice.onBlock` must never throw for a block that is valid with respect to the network
+   * `justifiedBalancesGetter()` must never throw and it should always return a state.
+   * @param blockState state that declares justified checkpoint `checkpoint`
+   */
+  private justifiedBalancesGetter(
+    checkpoint: CheckpointWithHex,
+    blockState: IBeaconStateView
+  ): EffectiveBalanceIncrements {
+    this.metrics?.balancesCache.requests.inc();
+
+    const effectiveBalances = this.checkpointBalancesCache.get(checkpoint);
+    if (effectiveBalances) {
+      return effectiveBalances;
+    }
+    // not expected, need metrics
+    this.metrics?.balancesCache.misses.inc();
+    this.logger.debug("checkpointBalances cache miss", {
+      epoch: checkpoint.epoch,
+      root: checkpoint.rootHex,
+    });
+
+    const {state, stateId, shouldWarn} = this.closestJustifiedBalancesStateToCheckpoint(checkpoint, blockState);
+    this.metrics?.balancesCache.closestStateResult.inc({stateId});
+    if (shouldWarn) {
+      this.logger.warn("currentJustifiedCheckpoint state not avail, using closest state", {
+        checkpointEpoch: checkpoint.epoch,
+        checkpointRoot: checkpoint.rootHex,
+        stateId,
+        stateSlot: state.slot,
+        stateRoot: toRootHex(state.hashTreeRoot()),
+      });
+    }
+
+    return state.getEffectiveBalanceIncrementsZeroInactive();
+  }
+
+  /**
+   * - Assumptions + invariant this function is based on:
+   * - Our cache can only persist X states at once to prevent OOM
+   * - Some old states (including to-be justified checkpoint) may / must be dropped from the cache
+   * - Thus, there is no guarantee that the state for a justified checkpoint will be available in the cache
+   * @param blockState state that declares justified checkpoint `checkpoint`
+   */
+  private closestJustifiedBalancesStateToCheckpoint(
+    checkpoint: CheckpointWithHex,
+    blockState: IBeaconStateView
+  ): {state: IBeaconStateView; stateId: string; shouldWarn: boolean} {
+    const checkpointHex = {epoch: checkpoint.epoch, rootHex: checkpoint.rootHex};
+    const state = this.regen.getCheckpointStateSync(checkpointHex);
+    if (state) {
+      return {state, stateId: "checkpoint_state", shouldWarn: false};
+    }
+
+    // Check if blockState is in the same epoch, not need to iterate the fork-choice then
+    if (computeEpochAtSlot(blockState.slot) === checkpoint.epoch) {
+      return {state: blockState, stateId: "block_state_same_epoch", shouldWarn: true};
+    }
+
+    // Find a state in the same branch of checkpoint at same epoch. Balances should exactly the same
+    for (const descendantBlock of this.forkChoice.forwardIterateDescendantsDefaultStatus(checkpoint.rootHex)) {
+      if (computeEpochAtSlot(descendantBlock.slot) === checkpoint.epoch) {
+        const descendantBlockState = this.regen.getStateSync(descendantBlock.stateRoot);
+        if (descendantBlockState) {
+          return {state: descendantBlockState, stateId: "descendant_state_same_epoch", shouldWarn: true};
+        }
+      }
+    }
+
+    // Check if blockState is in the next epoch, not need to iterate the fork-choice then
+    if (computeEpochAtSlot(blockState.slot) === checkpoint.epoch + 1) {
+      return {state: blockState, stateId: "block_state_next_epoch", shouldWarn: true};
+    }
+
+    // Find a state in the same branch of checkpoint at a latter epoch. Balances are not the same, but should be close
+    // Note: must call .forwardIterateDescendants() again since nodes are not sorted
+    for (const descendantBlock of this.forkChoice.forwardIterateDescendantsDefaultStatus(checkpoint.rootHex)) {
+      if (computeEpochAtSlot(descendantBlock.slot) > checkpoint.epoch) {
+        const descendantBlockState = this.regen.getStateSync(descendantBlock.stateRoot);
+        if (descendantBlockState) {
+          return {state: blockState, stateId: "descendant_state_latter_epoch", shouldWarn: true};
+        }
+      }
+    }
+
+    // If there's no state available in the same branch of checkpoint use blockState regardless of its epoch
+    return {state: blockState, stateId: "block_state_any_epoch", shouldWarn: true};
   }
 }
